Fix changing user experiment

+++ Modules/AbRouter/Services/Experiment/UserExperimentsService.php
@@ -95,7 +95,10 @@ class UserExperimentsService
                 $experimentUser->id
             );
 
-        $experimentBranchUser->delete();
+        if ($experimentBranchUser) {
+            $experimentBranchUser->delete();
+        }
+
         $experimentBranchUser = $this
             ->branchUserManager
             ->createExperimentBranchUser(
